% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnxFortran-symbols.R
\name{akima-symbols}
\alias{akima-symbols}
\alias{intrpl}
\alias{uvip3p}
\title{Akima Splines Fortran Interface}
\description{
Akima Splines Fortran Interface
}
\details{
For details see: \code{\link[=aspline-interface]{.aspline}}.
}
\section{Symbols}{
 \code{intrpl}, \code{uvip3p}
}

\seealso{
Other interfaces: 
\code{\link{bnxFortran-callable}},
\code{\link{bnxFortran-interface}},
\code{\link{nnls-symbols}}
}
\concept{interfaces}
\keyword{internal}
