% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnxFortran-package.R
\name{argCallby}
\alias{argCallby}
\title{Documentation of Recurring Package Arguments}
\arguments{
\item{callby}{A single character scalar defining the interface to the
exported function. This must be one of \code{"C"} indicates the \code{.C}
foreign function interface, \code{"Fortran"} indicates the \code{.Fortran}
foreign function interface, or \code{"Call"} indicates the \code{.Call}
modern C/C++ function interface. The supported \code{callby} choices
depend on the individual function. Any unambiguous substring can be given.}
}
\description{
Documentation of Recurring Package Arguments
}
\keyword{internal}
