% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnxFortran-callable.R
\name{BF_C_uvip3p}
\alias{BF_C_uvip3p}
\title{Akima Splines C Callable Interface: ACM 697}
\usage{
BF_C_uvip3p(np, n, x, y, rn, rx, ry, err)
}
\arguments{
\item{n}{An integer scalar from e.g. \code{as.integer()}. The number of input
data points. Must be 2 or greater.}

\item{x}{An double vector from e.g. \code{as.double()}. A numeric vector with
the x-values (abscissas) of input data points in ascending order.}

\item{y}{An double vector from e.g. \code{as.double()}. A numeric vector with
the y-values (ordinates) of input data points.}

\item{rn}{An integer scalar from e.g. \code{as.integer()}. The number of
points at which interpolation of the y-value (ordinate) is desired. Must be
1 or greater.}

\item{rx}{An double vector from e.g. \code{as.double()}. A vector of length
\code{rn} for the x-values (abscissas) of desired points.}

\item{ry}{An double vector from e.g. \code{as.double()}. A vector of length
\code{rn} where the interpolated y-values (ordinates) are to be displayed.}

\item{err}{An integer scalar from e.g. \code{as.integer()}. A variable to
store an error code.}

\item{np}{An R integer scalar from e.g. \code{as.integer()}. The degree of
the polynomials for the interpolating function.}
}
\description{
Akima Splines C Callable Interface: ACM 697
}
\details{
For details see: \code{\link[=aspline-interface]{.aspline}}.
}
\section{Symbols}{
 \code{BF_C_uvip3p}
}

\section{Callable}{
 The callable function is defined in a C header file as
follows: \cr \emph{int(\emph{BF_C_uvip3p)(int \emph{np, int \emph{n, double \emph{x, double \emph{y,
int \emph{rn, double \emph{rx, double \emph{ry, int \emph{err);} \cr\cr Within the
\code{R_init_...} function, the callable function is imported via \cr
\emph{BF_C_uvip3p = (int(})(int}, int}, double}, double}, int}, double},
double}, int})) R_GetCCallable("bnxFortran", "BF_C_uvip3p");}
}

\seealso{
Other callable: \code{\link{BF_C_intrpl}},
  \code{\link{BF_C_nnls}},
  \code{\link{bnxFortran-callable}}

Other akima-callable: \code{\link{BF_C_intrpl}}
}
\concept{akima-callable}
\concept{callable}
