% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnxFortran-callable.R
\name{bnxFortran-callable}
\alias{bnxFortran-callable}
\title{Callable Interface Function in \strong{bnxFortran} Package}
\description{
All callable functions can be called directly from other package source code.
For details see the details of respective functions listed below. Function
argument validation, except otherwise stated, shall usually be done by the
caller.
}
\seealso{
Other interfaces: \code{\link{akima-symbols}},
  \code{\link{bnxFortran-interface}},
  \code{\link{nnls-symbols}}

Other callable: \code{\link{BF_C_intrpl}},
  \code{\link{BF_C_nnls}}, \code{\link{BF_C_uvip3p}}
}
\concept{callable}
\concept{interfaces}
\keyword{internal}
