% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnxFortran-defunct.R
\name{bnxFortran-defunct}
\alias{bnxFortran-defunct}
\title{Defunct Functions in\strong{bnxFortran} Package}
\arguments{
\item{...}{any argument to be passed down.}
}
\description{
The functions or variables listed here are no longer part of the package as
they are no longer needed or backward compatibility is not anymore possible.
}
\details{
Some of these have stubs which report that they are defunct, but
most have been removed completely (apart from being documented here).
}
\seealso{
\code{\link[base]{Defunct}}
}
\keyword{internal}
