% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnxFortran-deprecated.R
\name{bnxFortran-deprecated}
\alias{bnxFortran-deprecated}
\title{Deprecated Functions in\strong{bnxFortran} Package}
\arguments{
\item{...}{any argument to be passed down.}
}
\description{
These functions are provided for compatibility with older versions of the
package only, and may be defunct as soon as the next release or at least in
the near future. Commands that worked in previous versions of this package
may not necessarily work in the current version and beyond, or may not work
in the same manner.
}
\seealso{
\code{\link[base]{Deprecated}}
}
\keyword{internal}
