% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnxFortran-symbols.R
\name{bnxFortran-interface}
\alias{bnxFortran-interface}
\title{Interface Symbols in \strong{bnxFortran} Package}
\description{
All functions referenced by the given hyperlinks are external interface
functions used within this package. These functions shall not be called
directly as their interface may change in future releases.
}
\seealso{
Other interfaces: \code{\link{akima-symbols}},
  \code{\link{bnxFortran-callable}},
  \code{\link{nnls-symbols}}
}
\concept{interfaces}
\keyword{internal}
