% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnxFortran-package.R
\name{bnxFortran}
\alias{bnxFortran}
\title{bnxFortran: BNX Interface Package for R2C Callable Fortran Routines}
\description{
The 'bnxFortran' package provides some C functions that wrap Fortran
algorithms. These interface functions are implemented here to be called
directly from the C or C++ source code level of other R packages.
}
\details{
\bold{No S4 methodology allowed here!}

See documentation for overview and details.
}
\references{
See documentation for overview and details.
}
\author{
Dirk Steinhauser <\email{steinhauser@bionumerix.org}>

See \code{DESCRIPTION} and \code{README.md} for contributors.
}
\keyword{package}
