% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnxFortran-symbols.R
\name{nnls-symbols}
\alias{nnls-symbols}
\alias{nnls}
\title{Non-Negative Least Squares (NNLS) Fortran Interface}
\description{
Non-Negative Least Squares (NNLS) Fortran Interface
}
\details{
For details see: \code{\link[=nnls-interface]{.nnls}}.
}
\section{Symbols}{
 \code{nnls}
}

\seealso{
Other interfaces: 
\code{\link{akima-symbols}},
\code{\link{bnxFortran-callable}},
\code{\link{bnxFortran-interface}}
}
\concept{interfaces}
\keyword{internal}
